Kubernetes version v0.9.2

Conflicts:
	Godeps/Godeps.json
	README.md
	Vagrantfile
	cluster/kube-env.sh
	cluster/saltbase/salt/docker/docker-defaults
	cluster/saltbase/salt/docker/init.sls
	cluster/saltbase/salt/kube-scheduler/default
	cluster/vagrant/config-default.sh
	pkg/version/base.go
	plugin/cmd/kube-scheduler/scheduler.go
	plugin/pkg/scheduler/factory/factory.go

+++ build/common.sh
@@ -34,14 +34,14 @@ source hack/lib/init.sh
 # Incoming options
 #
 readonly KUBE_SKIP_CONFIRMATIONS="${KUBE_SKIP_CONFIRMATIONS:-n}"
-readonly KUBE_BUILD_RUN_IMAGES="${KUBE_BUILD_RUN_IMAGES:-n}"
 readonly KUBE_GCS_UPLOAD_RELEASE="${KUBE_GCS_UPLOAD_RELEASE:-n}"
 readonly KUBE_GCS_NO_CACHING="${KUBE_GCS_NO_CACHING:-y}"
 readonly KUBE_GCS_MAKE_PUBLIC="${KUBE_GCS_MAKE_PUBLIC:-y}"
 # KUBE_GCS_RELEASE_BUCKET default: kubernetes-releases-${project_hash}
 readonly KUBE_GCS_RELEASE_PREFIX=${KUBE_GCS_RELEASE_PREFIX-devel}/
 readonly KUBE_GCS_DOCKER_REG_PREFIX=${KUBE_GCS_DOCKER_REG_PREFIX-docker-reg}/
-
+readonly KUBE_GCS_LATEST_FILE=${KUBE_GCS_LATEST_FILE:-}
+readonly KUBE_GCS_LATEST_CONTENTS=${KUBE_GCS_LATEST_CONTENTS:-}
 
 # Constants
 readonly KUBE_BUILD_IMAGE_REPO=kube-build
@@ -86,18 +86,6 @@ readonly DOCKER_DATA_MOUNT_ARGS=(
   --volume "${REMOTE_GODEP_GOPATH}"
 )
 
-
-readonly KUBE_RUN_IMAGE_BASE="kubernetes"
-readonly KUBE_RUN_IMAGES=(
-  kube-apiserver
-  kube-controller-manager
-  kube-proxy
-  kube-scheduler
-  kubelet
-  bootstrap
-)
-
-
 # This is where the final release artifacts are created locally
 readonly RELEASE_STAGE="${LOCAL_OUTPUT_ROOT}/release-stage"
 readonly RELEASE_DIR="${LOCAL_OUTPUT_ROOT}/release-tars"
@@ -108,6 +96,10 @@ readonly RELEASE_DIR="${LOCAL_OUTPUT_ROOT}/release-tars"
 # Verify that the right utilities and such are installed for building Kube.  Set
 # up some dynamic constants.
 #
+# Args:
+#   $1 The type of operation to verify for.  Only 'clean' is supported in which
+#   case we don't verify docker.
+#
 # Vars set:
 #   KUBE_ROOT_HASH
 #   KUBE_BUILD_IMAGE_TAG
@@ -117,44 +109,47 @@ readonly RELEASE_DIR="${LOCAL_OUTPUT_ROOT}/release-tars"
 #   DOCKER_MOUNT_ARGS
 function kube::build::verify_prereqs() {
   echo "+++ Verifying Prerequisites...."
-  if [[ -z "$(which docker)" ]]; then
-    echo "Can't find 'docker' in PATH, please fix and retry." >&2
-    echo "See https://docs.docker.com/installation/#installation for installation instructions." >&2
-    exit 1
-  fi
 
-  if kube::build::is_osx; then
-    if [[ -z "$DOCKER_NATIVE" ]];then
-      if [[ -z "$(which boot2docker)" ]]; then
-        echo "It looks like you are running on Mac OS X and boot2docker can't be found." >&2
-        echo "See: https://docs.docker.com/installation/mac/" >&2
-        exit 1
-      fi
-      if [[ $(boot2docker status) != "running" ]]; then
-        echo "boot2docker VM isn't started.  Please run 'boot2docker start'" >&2
-        exit 1
-      else
-        # Reach over and set the clock. After sleep/resume the clock will skew.
-        echo "+++ Setting boot2docker clock"
-        boot2docker ssh sudo date -u -D "%Y%m%d%H%M.%S" --set "$(date -u +%Y%m%d%H%M.%S)" >/dev/null
+  if [[ "${1-}" != "clean" ]]; then
+    if [[ -z "$(which docker)" ]]; then
+      echo "Can't find 'docker' in PATH, please fix and retry." >&2
+      echo "See https://docs.docker.com/installation/#installation for installation instructions." >&2
+      exit 1
+    fi
+
+    if kube::build::is_osx; then
+      if [[ -z "$DOCKER_NATIVE" ]];then
+        if [[ -z "$(which boot2docker)" ]]; then
+          echo "It looks like you are running on Mac OS X and boot2docker can't be found." >&2
+          echo "See: https://docs.docker.com/installation/mac/" >&2
+          exit 1
+        fi
+        if [[ $(boot2docker status) != "running" ]]; then
+          echo "boot2docker VM isn't started.  Please run 'boot2docker start'" >&2
+          exit 1
+        else
+          # Reach over and set the clock. After sleep/resume the clock will skew.
+          echo "+++ Setting boot2docker clock"
+          boot2docker ssh sudo date -u -D "%Y%m%d%H%M.%S" --set "$(date -u +%Y%m%d%H%M.%S)" >/dev/null
+        fi
       fi
     fi
-  fi
 
-  if ! "${DOCKER[@]}" info > /dev/null 2>&1 ; then
-    {
-      echo "Can't connect to 'docker' daemon.  please fix and retry."
-      echo
-      echo "Possible causes:"
-      echo "  - On Mac OS X, boot2docker VM isn't installed or started"
-      echo "  - On Mac OS X, docker env variable isn't set approriately. Run:"
-      echo "      \$(boot2docker shellinit)"
-      echo "  - On Linux, user isn't in 'docker' group.  Add and relogin."
-      echo "    - Something like 'sudo usermod -a -G docker ${USER-user}'"
-      echo "    - RHEL7 bug and workaround: https://bugzilla.redhat.com/show_bug.cgi?id=1119282#c8"
-      echo "  - On Linux, Docker daemon hasn't been started or has crashed"
-    } >&2
-    exit 1
+    if ! "${DOCKER[@]}" info > /dev/null 2>&1 ; then
+      {
+        echo "Can't connect to 'docker' daemon.  please fix and retry."
+        echo
+        echo "Possible causes:"
+        echo "  - On Mac OS X, boot2docker VM isn't installed or started"
+        echo "  - On Mac OS X, docker env variable isn't set appropriately. Run:"
+        echo "      \$(boot2docker shellinit)"
+        echo "  - On Linux, user isn't in 'docker' group.  Add and relogin."
+        echo "    - Something like 'sudo usermod -a -G docker ${USER-user}'"
+        echo "    - RHEL7 bug and workaround: https://bugzilla.redhat.com/show_bug.cgi?id=1119282#c8"
+        echo "  - On Linux, Docker daemon hasn't been started or has crashed"
+      } >&2
+      exit 1
+    fi
   fi
 
   KUBE_ROOT_HASH=$(kube::build::short_hash "$KUBE_ROOT")
@@ -174,15 +169,17 @@ function kube::build::is_osx() {
 
 function kube::build::clean_output() {
   # Clean out the output directory if it exists.
-  if kube::build::build_image_built ; then
-    echo "+++ Cleaning out _output/dockerized/bin/ via docker build image"
-    kube::build::run_build_command bash -c "rm -rf '${REMOTE_OUTPUT_BINPATH}'/*"
-  else
-    echo "!!! Build image not built.  Cannot clean via docker build image."
-  fi
+  if kube::build::has_docker ; then
+    if kube::build::build_image_built ; then
+      echo "+++ Cleaning out _output/dockerized/bin/ via docker build image"
+      kube::build::run_build_command bash -c "rm -rf '${REMOTE_OUTPUT_BINPATH}'/*"
+    else
+      echo "!!! Build image not built.  Cannot clean via docker build image."
+    fi
 
-  echo "+++ Removing data container"
-  "${DOCKER[@]}" rm -v "${KUBE_BUILD_DATA_CONTAINER_NAME}" >/dev/null 2>&1 || true
+    echo "+++ Removing data container"
+    "${DOCKER[@]}" rm -v "${KUBE_BUILD_DATA_CONTAINER_NAME}" >/dev/null 2>&1 || true
+  fi
 
   echo "+++ Cleaning out local _output directory"
   rm -rf "${LOCAL_OUTPUT_ROOT}"
@@ -208,6 +205,10 @@ function kube::build::prepare_output() {
     fi
   fi
 
+}
+
+function kube::build::has_docker() {
+  which docker &> /dev/null
 }
 
 # Detect if a specific image exists
@@ -240,6 +241,17 @@ function kube::build::short_hash() {
   fi
   echo ${short_hash:0:5}
 }
+
+# Pedantically kill, wait-on and remove a container. The -f -v options
+# to rm don't actually seem to get the job done, so force kill the
+# container, wait to ensure it's stopped, then try the remove. This is
+# a workaround for bug https://github.com/docker/docker/issues/3968.
+function kube::build::destroy_container() {
+  "${DOCKER[@]}" kill "$1" >/dev/null 2>&1 || true
+  "${DOCKER[@]}" wait "$1" >/dev/null 2>&1 || true
+  "${DOCKER[@]}" rm -f -v "$1" >/dev/null 2>&1 || true
+}
+
 
 # ---------------------------------------------------------------------------
 # Building
@@ -306,30 +318,6 @@ function kube::build::build_image_cross() {
   kube::build::docker_build "${KUBE_BUILD_IMAGE_CROSS}" "${build_context_dir}"
 }
 
-# Builds the runtime image.  Assumes that the appropriate binaries are already
-# built and in $LOCAL_OUTPUT_BINPATH.
-function kube::build::run_image() {
-  [[ ${KUBE_BUILD_RUN_IMAGES} =~ ^[yY]$ ]] || return 0
-
-  local -r build_context_base="${LOCAL_OUTPUT_IMAGE_STAGING}/${KUBE_RUN_IMAGE_BASE}"
-
-  # First build the base image.  This one brings in all of the binaries.
-  mkdir -p "${build_context_base}"
-  tar czf "${build_context_base}/kube-bins.tar.gz" \
-    -C "${LOCAL_OUTPUT_BINPATH}/linux/amd64" \
-    "${KUBE_RUN_IMAGES[@]}"
-  cp -R build/run-images/base/* "${build_context_base}/"
-  kube::build::docker_build "${KUBE_RUN_IMAGE_BASE}" "${build_context_base}"
-
-  local b
-  for b in "${KUBE_RUN_IMAGES[@]}" ; do
-    local sub_context_dir="${build_context_base}-$b"
-    mkdir -p "${sub_context_dir}"
-    cp -R build/run-images/$b/* "${sub_context_dir}/"
-    kube::build::docker_build "${KUBE_RUN_IMAGE_BASE}-$b" "${sub_context_dir}"
-  done
-}
-
 # Build a docker image from a Dockerfile.
 # $1 is the name of the image to build
 # $2 is the location of the "context" directory, with the Dockerfile at the root.
@@ -363,14 +351,9 @@ function kube::build::clean_image() {
 }
 
 function kube::build::clean_images() {
+  kube::build::has_docker || return 0
+
   kube::build::clean_image "${KUBE_BUILD_IMAGE}"
-
-  kube::build::clean_image "${KUBE_RUN_IMAGE_BASE}"
-
-  local b
-  for b in "${KUBE_RUN_IMAGES[@]}" ; do
-    kube::build::clean_image "${KUBE_RUN_IMAGE_BASE}-${b}"
-  done
 
   echo "+++ Cleaning all other untagged docker images"
   "${DOCKER[@]}" rmi $("${DOCKER[@]}" images -q --filter 'dangling=true') 2> /dev/null || true
@@ -417,14 +400,10 @@ function kube::build::run_build_command() {
   local -ra docker_cmd=(
     "${DOCKER[@]}" run "${docker_run_opts[@]}" "${KUBE_BUILD_IMAGE}")
 
-  # Remove the container if it is left over from some previous aborted run
-  "${DOCKER[@]}" rm -f -v "${KUBE_BUILD_CONTAINER_NAME}" >/dev/null 2>&1 || true
+  # Clean up container from any previous run
+  kube::build::destroy_container "${KUBE_BUILD_CONTAINER_NAME}"
   "${docker_cmd[@]}" "$@"
-
-  # Remove the container after we run.  '--rm' might be appropriate but it
-  # appears that sometimes it fails. See
-  # https://github.com/docker/docker/issues/3968
-  "${DOCKER[@]}" rm -f -v "${KUBE_BUILD_CONTAINER_NAME}" >/dev/null 2>&1 || true
+  kube::build::destroy_container "${KUBE_BUILD_CONTAINER_NAME}"
 }
 
 # Test if the output directory is remote (and can only be accessed through
@@ -460,8 +439,7 @@ function kube::build::copy_output() {
     rm -rf "${LOCAL_OUTPUT_BINPATH}"
     mkdir -p "${LOCAL_OUTPUT_BINPATH}"
 
-    # Remove the container if it is left over from some previous aborted run
-    "${DOCKER[@]}" rm -f -v "${KUBE_BUILD_CONTAINER_NAME}" >/dev/null 2>&1 || true
+    kube::build::destroy_container "${KUBE_BUILD_CONTAINER_NAME}"
     "${docker_cmd[@]}" bash -c "cp -r ${REMOTE_OUTPUT_BINPATH} /tmp/bin;touch /tmp/finished;rm /tmp/bin/test_for_remote;/bin/sleep 600" > /dev/null 2>&1
 
     # Wait until binaries have finished coppying
@@ -497,6 +475,7 @@ function kube::release::package_tarballs() {
   kube::release::package_client_tarballs
   kube::release::package_server_tarballs
   kube::release::package_salt_tarball
+  kube::release::package_test_tarball
   kube::release::package_full_tarball
 }
 
@@ -572,7 +551,40 @@ function kube::release::package_salt_tarball() {
 
   cp -R "${KUBE_ROOT}/cluster/saltbase" "${release_stage}/"
 
+  # TODO(#3579): This is a temporary hack. It gathers up the yaml,
+  # yaml.in files in cluster/addons (minus any demos) and overlays
+  # them into kube-addons, where we expect them. (This pipeline is a
+  # fancy copy, stripping anything but the files we don't want.)
+  local objects
+  objects=$(cd "${KUBE_ROOT}/cluster/addons" && find . -name \*.yaml -or -name \*.yaml.in | grep -v demo)
+  tar c -C "${KUBE_ROOT}/cluster/addons" ${objects} | tar x -C "${release_stage}/saltbase/salt/kube-addons"
+
   local package_name="${RELEASE_DIR}/kubernetes-salt.tar.gz"
+  kube::release::create_tarball "${package_name}" "${release_stage}/.."
+}
+
+# This is the stuff you need to run tests from the binary distribution.
+function kube::release::package_test_tarball() {
+  echo "+++ Building tarball: test"
+
+  local release_stage="${RELEASE_STAGE}/test/kubernetes"
+  rm -rf "${release_stage}"
+  mkdir -p "${release_stage}"
+
+  local platform
+  for platform in "${KUBE_CLIENT_PLATFORMS[@]}"; do
+    local test_bins=("${KUBE_TEST_BINARIES[@]}")
+    if [[ "${platform%/*}" == "windows" ]]; then
+      test_bins=("${KUBE_TEST_BINARIES_WIN[@]}")
+    fi
+    mkdir -p "${release_stage}/platforms/${platform}"
+    cp "${test_bins[@]/#/${LOCAL_OUTPUT_BINPATH}/${platform}/}" \
+      "${release_stage}/platforms/${platform}"
+  done
+
+  tar c ${KUBE_TEST_PORTABLE[@]} | tar x -C ${release_stage}
+
+  local package_name="${RELEASE_DIR}/kubernetes-test.tar.gz"
   kube::release::create_tarball "${package_name}" "${release_stage}/.."
 }
 
@@ -655,7 +667,6 @@ function kube::release::gcs::release() {
 
   kube::release::gcs::verify_prereqs
   kube::release::gcs::ensure_release_bucket
-  kube::release::gcs::push_images
   kube::release::gcs::copy_release_artifacts
 }
 
@@ -700,59 +711,6 @@ function kube::release::gcs::ensure_release_bucket() {
     echo "Creating Google Cloud Storage bucket: $KUBE_GCS_RELEASE_BUCKET"
     gsutil mb -p "${GCLOUD_PROJECT}" "gs://${KUBE_GCS_RELEASE_BUCKET}"
   fi
-}
-
-function kube::release::gcs::ensure_docker_registry() {
-  local -r reg_container_name="gcs-registry"
-
-  local -r running=$("${DOCKER[@]}" inspect ${reg_container_name} 2>/dev/null \
-    | build/json-extractor.py 0.State.Running 2>/dev/null)
-
-  [[ "$running" != "true" ]] || return 0
-
-  # Grovel around and find the OAuth token in the gcloud config
-  local -r boto=~/.config/gcloud/legacy_credentials/${GCLOUD_ACCOUNT}/.boto
-  local refresh_token
-  refresh_token=$(grep 'gs_oauth2_refresh_token =' "$boto" | awk '{ print $3 }')
-
-  if [[ -z "$refresh_token" ]]; then
-    echo "Couldn't find OAuth 2 refresh token in ${boto}" >&2
-    return 1
-  fi
-
-  # If we have an old one sitting around, remove it
-  "${DOCKER[@]}" rm ${reg_container_name} >/dev/null 2>&1 || true
-
-  echo "+++ Starting GCS backed Docker registry"
-  local -ra docker_cmd=(
-    "${DOCKER[@]}" run -d "--name=${reg_container_name}"
-    -e "GCS_BUCKET=${KUBE_GCS_RELEASE_BUCKET}"
-    -e "STORAGE_PATH=${KUBE_GCS_DOCKER_REG_PREFIX}"
-    -e "GCP_OAUTH2_REFRESH_TOKEN=${refresh_token}"
-    -p 127.0.0.1:5000:5000
-    google/docker-registry
-  )
-
-  "${docker[@]}"
-
-  # Give it time to spin up before we start throwing stuff at it
-  sleep 5
-}
-
-function kube::release::gcs::push_images() {
-  [[ ${KUBE_BUILD_RUN_IMAGES} =~ ^[yY]$ ]] || return 0
-
-  kube::release::gcs::ensure_docker_registry
-
-  # Tag each of our run binaries with the right registry and push
-  local b image_name
-  for b in "${KUBE_RUN_IMAGES[@]}" ; do
-    image_name="${KUBE_RUN_IMAGE_BASE}-${b}"
-    echo "+++ Tagging and pushing ${image_name} to GCS bucket ${KUBE_GCS_RELEASE_BUCKET}"
-    "${DOCKER[@]}" tag "${KUBE_RUN_IMAGE_BASE}-$b" "localhost:5000/${image_name}"
-    "${DOCKER[@]}" push "localhost:5000/${image_name}"
-    "${DOCKER[@]}" rmi "localhost:5000/${image_name}"
-  done
 }
 
 function kube::release::gcs::copy_release_artifacts() {
@@ -818,4 +776,21 @@ function kube::release::gcs::copy_release_artifacts() {
   fi
 
   gsutil ls -lhr "${gcs_destination}"
+}
+
+function kube::release::gcs::publish_latest() {
+  local latest_file_dst="gs://${KUBE_GCS_RELEASE_BUCKET}/${KUBE_GCS_LATEST_FILE}"
+
+  mkdir -p "${RELEASE_STAGE}/upload"
+  echo ${KUBE_GCS_LATEST_CONTENTS} > "${RELEASE_STAGE}/upload/latest"
+
+  gsutil -m "${gcs_options[@]+${gcs_options[@]}}" cp \
+    "${RELEASE_STAGE}/upload/latest" "${latest_file_dst}"
+
+  if [[ ${KUBE_GCS_MAKE_PUBLIC} =~ ^[yY]$ ]]; then
+    gsutil acl ch -R -g all:R "${latest_file_dst}" >/dev/null 2>&1
+  fi
+
+  echo "+++ gsutil cat ${latest_file_dst}:"
+  gsutil cat ${latest_file_dst}
 }
